Moved settings to a nested control in main window
