+++ XIVLauncher/Properties/Resources.resx
@@ -118,8 +118,29 @@
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
   <assembly alias="System.Windows.Forms" name="System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
+  <data name="about_invert" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\about_invert.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="cactuar_invert" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\cactuar_invert.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="chocobo_invert" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\chocobo_invert.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
   <data name="dalamud_icon" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\dalamud_icon.ico;System.Drawing.Icon, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="Discord_Logo_invert" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\Discord-Logo-invert.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="gear_invert" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\gear_invert.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="github_invert2" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\github_invert2.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="guide_invert" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\guide_invert.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
   <assembly alias="System.Drawing" name="System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a" />
   <data name="placeholder_news" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
@@ -4261,5 +4282,17 @@
         h3WOsv8fW1m8MVnlbzRcv1i+rN+WU4K1/o3VxZvTW3XcYVn/4c9LTtdEkYHOjE/LpMs3qPNHbfegMDtJ
         HGB+F18eHO/vrYrHdlo+rMPXc2Odkv8fRdsxD8iMfSIAAAAASUVORK5CYII=
 </value>
+  </data>
+  <data name="scroll_invert" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\scroll_invert.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="world_invert" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\world_invert.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="world_red" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\world_red.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="_340" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\340.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
 </root>
