+++ XIVLauncher/packages.config
@@ -1,17 +1,17 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="AdysTech.CredentialManager" version="1.8.0" targetFramework="net461" />
-  <package id="Autoupdater.NET.Official" version="1.5.1" targetFramework="net461" />
-  <package id="Costura.Fody" version="3.3.2" targetFramework="net461" />
+  <package id="Autoupdater.NET.Official" version="1.5.1" targetFramework="net462" />
+  <package id="Costura.Fody" version="3.3.2" targetFramework="net462" />
   <package id="DeltaCompressionDotNet" version="2.0.0.0" targetFramework="net461" />
   <package id="Dragablz" version="0.0.3.203" targetFramework="net461" />
   <package id="Extended.Wpf.Toolkit" version="3.5.0" targetFramework="net462" />
-  <package id="Fody" version="3.3.5" targetFramework="net461" developmentDependency="true" />
+  <package id="Fody" version="3.3.5" targetFramework="net462" developmentDependency="true" />
   <package id="MaterialDesignColors" version="1.1.2" targetFramework="net462" />
   <package id="MaterialDesignThemes" version="2.5.1" targetFramework="net462" />
-  <package id="Microsoft.WindowsAPICodePack-Core" version="1.1.0.0" targetFramework="net461" />
+  <package id="Microsoft.WindowsAPICodePack-Core" version="1.1.0.0" targetFramework="net462" />
   <package id="Microsoft.WindowsAPICodePack-Shell" version="1.1.0.0" targetFramework="net461" />
-  <package id="Newtonsoft.Json" version="12.0.1" targetFramework="net461" />
+  <package id="Newtonsoft.Json" version="12.0.1" targetFramework="net462" />
   <package id="Serilog" version="2.8.0" targetFramework="net462" />
   <package id="Serilog.Sinks.Async" version="1.4.0" targetFramework="net462" />
   <package id="Serilog.Sinks.File" version="4.0.0" targetFramework="net462" />
